requirements.txt
python-snap7
numpy
PySimpleGUI
